+++ src/imio/fpaudit/subscribers.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from imio.fpaudit import LOG_ENTRIES_REGISTRY
 from imio.fpaudit.storage import store_config
 from imio.helpers.security import get_zope_root
 from imio.helpers.security import set_site_from_package_config
@@ -18,7 +19,7 @@ def zope_ready(event):
             "/instance"
         ):
             with api.env.adopt_user("admin"):
-                log_entries = api.portal.get_registry_record("imio.fpaudit.settings.log_entries", default=[])
+                log_entries = api.portal.get_registry_record(LOG_ENTRIES_REGISTRY, default=[])
                 if log_entries:
                     store_config(log_entries)
                     change = True